+++ Infrastructure/Presentation/OtherController.cs
@@ -1,5 +1,6 @@
 ﻿using Apple1_Services.Abstractions;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using Shared;
 using System;
 using System.Collections.Generic;
@@ -11,8 +12,15 @@ namespace Presentation
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class OtherController(IServiceManager serviceManager) : ControllerBase
+    public class OtherController : ControllerBase
     {
+        private readonly string _adminPassword;
+        private readonly IServiceManager serviceManager;
+        public OtherController(IOptions<AdminSettings> adminSettingsOptions, IServiceManager service)
+        {
+            _adminPassword = adminSettingsOptions.Value.Password;
+            serviceManager = service;
+        }
         [HttpGet]
         public async Task<IActionResult> GetAllOthers()
         {
@@ -22,7 +30,7 @@ namespace Presentation
         }
 
         [HttpGet("{Name}")]
-        public async Task<IActionResult> GetOtherByName([FromQuery] string name)
+        public async Task<IActionResult> GetOtherByName(string name)
         {
             var other = await serviceManager.OthersService.GetOtherByNameAsync(name);
             if (other == null) return NotFound($"Other item with name '{name}' not found.");
@@ -30,24 +38,34 @@ namespace Presentation
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateOther([FromBody] AddOthersResultDto otherDto)
+        public async Task<IActionResult> CreateOther([FromBody] AddOthersResultDto otherDto, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
+            if (ModelState.IsValid == false)
+            {
+                return BadRequest(ModelState);
+            }
             if (otherDto == null) return BadRequest("Other item data is null.");
             await serviceManager.OthersService.CreateOtherAsync(otherDto);
             return Ok(otherDto);
         }
 
         [HttpPut]
-        public async Task<IActionResult> UpdateOther([FromBody] AddOthersResultDto otherDto)
+        public async Task<IActionResult> UpdateOther([FromBody] OthersResultDto otherDto, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
             if (otherDto == null) return BadRequest("Other item data is null.");
             await serviceManager.OthersService.UpdateOtherAsync(otherDto);
             return Ok(otherDto);
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteOther(int id)
+        public async Task<IActionResult> DeleteOther(int id, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
             await serviceManager.OthersService.DeleteOtherAsync(id);
             return Ok($"Other item with ID '{id}' has been deleted.");
         }
