+++ Infrastructure/Presentation/ScreenController.cs
@@ -1,5 +1,6 @@
 ﻿using Apple1_Services.Abstractions;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using Shared;
 using System;
 using System.Collections.Generic;
@@ -11,8 +12,15 @@ namespace Presentation
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class ScreenController(IServiceManager serviceManager) : ControllerBase
+    public class ScreenController : ControllerBase
     {
+        private readonly string _adminPassword;
+        private readonly IServiceManager serviceManager;
+        public ScreenController(IOptions<AdminSettings> adminSettingsOptions, IServiceManager service)
+        {
+            _adminPassword = adminSettingsOptions.Value.Password;
+            serviceManager = service;
+        }
         [HttpGet]
         public async Task<IActionResult> GetAllScreens()
         {
@@ -22,7 +30,7 @@ namespace Presentation
         }
 
         [HttpGet("{Name}")]
-        public async Task<IActionResult> GetScreenByName([FromQuery] string name)
+        public async Task<IActionResult> GetScreenByName( string name)
         {
             var screen = await serviceManager.ScreenService.GetScreenByNameAsync(name);
             if (screen == null) return NotFound($"Screen with name '{name}' not found.");
@@ -30,24 +38,34 @@ namespace Presentation
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateScreen([FromBody] AddScreenResultDto screenDto)
+        public async Task<IActionResult> CreateScreen([FromBody] AddScreenResultDto screenDto, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
+            if (ModelState.IsValid == false)
+            {
+                return BadRequest(ModelState);
+            }
             if (screenDto == null) return BadRequest("Screen data is null.");
             await serviceManager.ScreenService.CreateScreenAsync(screenDto);
             return Ok(screenDto);
         }
 
         [HttpPut]
-        public async Task<IActionResult> UpdateScreen([FromBody] AddScreenResultDto screenDto)
+        public async Task<IActionResult> UpdateScreen([FromBody] ScreenResultDto screenDto, [FromHeader(Name = "Admin-Password")] string password)
         {
-            if (screenDto == null) return BadRequest("Screen data is null.");
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
+            if (screenDto == null ) return BadRequest("Screen data is null or Not Found.");
             await serviceManager.ScreenService.UpdateScreenAsync(screenDto);
             return Ok(screenDto);
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteScreen(int id)
+        public async Task<IActionResult> DeleteScreen(int id, [FromHeader(Name = "Admin-Password")] string password)
         {
+            if (password != _adminPassword)
+                return StatusCode(401, "Invalid password");
             await serviceManager.ScreenService.DeleteScreenAsync(id);
             return Ok($"Screen with ID '{id}' has been deleted.");
         }
